% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_predictors.R
\name{add_one_predictor_strings}
\alias{add_one_predictor_strings}
\title{Helper function that adds one predictor to a model sans quoting.}
\usage{
add_one_predictor_strings(df, my_formula, new_predictor)
}
\arguments{
\item{df}{A dataframe}

\item{my_formula}{A formula or string.}

\item{new_predictor}{An (QUOTED) variable to try in the model.}
}
\value{
A new model object
}
\description{
Helper function that adds one predictor to a model sans quoting.
}
