% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_SE_lm.R
\name{bootstrap_EST_df}
\alias{bootstrap_EST_df}
\title{Helper function that creates a dataframe of bootstrapped coefficient estimates given a regression, allowing
the user to create confidence intervals, calculate means, plot graphs, etc.}
\usage{
bootstrap_EST_df(df, response, predictors, number_of_reps)
}
\arguments{
\item{df}{A dataframe R will pull the variables from to run the regression}

\item{response}{The dependent variable of the regression, as a string}

\item{predictors}{A vector of strings containing each independent variable used in the model}

\item{number_of_reps}{Number of samples to take from the df and regressions to run}
}
\value{
A dataframe with a column for each variable with rows containing each sample regression's
estimates
}
\description{
Helper function that creates a dataframe of bootstrapped coefficient estimates given a regression, allowing
the user to create confidence intervals, calculate means, plot graphs, etc.
}
