% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_lm.R
\name{bootstrap_lm}
\alias{bootstrap_lm}
\title{Creates a dataframe of bootstrapped standard errors given a regression, allowing the user to create
confidence intervals, calculate means, plot graphs, etc.}
\usage{
bootstrap_lm(df, model, n = 1000)
}
\arguments{
\item{df}{A dataframe R will pull the variables from to run the regression}

\item{model}{A model object of class "lm" or a valid formula.}

\item{n}{Number of samples to take from the df and regressions to run}
}
\value{
A list of three data frames with bootstrapped estimates of coefficients,
standard errors, and summary statistics.
}
\description{
Creates a dataframe of bootstrapped standard errors given a regression, allowing the user to create
confidence intervals, calculate means, plot graphs, etc.
}
\examples{

bootstrap_lm(mtcars, mpg ~ cyl + am + hp)

}
