% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_SE_df.R
\name{bootstrap_SE_df}
\alias{bootstrap_SE_df}
\title{Creates a dataframe of bootstrapped standard errors given a regression, allowing the user to create
confidence intervals, calculate means, plot graphs, etc.}
\usage{
bootstrap_SE_df(df, response, predictors, number_of_reps)
}
\arguments{
\item{df}{A dataframe R will pull the variables from to run the regression}

\item{response}{The dependent variable of the regression, as a string}

\item{predictors}{A vector of strings containing each independent variable used in the model}

\item{number_of_reps}{Number of samples to take from the df and regressions to run}
}
\value{
A dataframe with a column for each variable with rows containing each sample regression's
standard errors
}
\description{
Creates a dataframe of bootstrapped standard errors given a regression, allowing the user to create
confidence intervals, calculate means, plot graphs, etc.
}
