% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_SE_lm.R
\name{bootstrap_SE_lm}
\alias{bootstrap_SE_lm}
\title{Runs a linear regression using bootstrapped estimates to create the covarience matrix.}
\usage{
bootstrap_SE_lm(df, response, predictors, number_of_reps)
}
\arguments{
\item{df}{A dataframe R will pull the variables from to run the regression}

\item{response}{The dependent variable of the regression, as a string}

\item{predictors}{A vector of strings containing each independent variable used in the model}

\item{number_of_reps}{Number of samples to take from the df and regressions to run}
}
\value{
A linear regression result, similar to summary of a model
}
\description{
Runs a linear regression using bootstrapped estimates to create the covarience matrix.
}
