% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R3.R
\name{bootstrapESTdf}
\alias{bootstrapESTdf}
\title{Creates a dataframe of bootstrapped coefficient estimates given a regression, allowing the user
to create confidence intervals, calculate means, plot graphs, etc.}
\usage{
bootstrapESTdf(df, output, startersvec, numreg)
}
\arguments{
\item{df}{A dataframe R will pull the variables from to run the regression}

\item{output}{The dependent variable of the regression, as a string}

\item{startersvec}{A vector of strings containing each independent variable used in the model}

\item{numreg}{Number of samples to take from the df and regressions to run}
}
\value{
A dataframe with a column for each variable with rows containing each sample regression's
estimates
}
\description{
Creates a dataframe of bootstrapped coefficient estimates given a regression, allowing the user
to create confidence intervals, calculate means, plot graphs, etc.
}
