% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R3.R
\name{bootstrapSEdf}
\alias{bootstrapSEdf}
\title{Creates a dataframe of bootstrapped standard errors given a regression, allowing the user to create
confidence intervals, calculate means, plot graphs, etc.}
\usage{
bootstrapSEdf(df, output, startersvec, numreg)
}
\arguments{
\item{df}{A dataframe R will pull the variables from to run the regression}

\item{output}{The dependent variable of the regression, as a string}

\item{startersvec}{A vector of strings containing each independent variable used in the model}

\item{numreg}{Number of samples to take from the df and regressions to run}
}
\value{
A dataframe with a column for each variable with rows containing each sample regression's
standard errors
}
\description{
Creates a dataframe of bootstrapped standard errors given a regression, allowing the user to create
confidence intervals, calculate means, plot graphs, etc.
}
