% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R3.R
\name{bootstrapSElm}
\alias{bootstrapSElm}
\title{Runs a linear regression using bootstrapped estimates to create the covarience matrix.}
\usage{
bootstrapSElm(df, output, startersvec, numreg)
}
\arguments{
\item{df}{A dataframe R will pull the variables from to run the regression}

\item{output}{The dependent variable of the regression, as a string}

\item{startersvec}{A vector of strings containing each independent variable used in the model}

\item{numreg}{Number of samples to take from the df and regressions to run}
}
\value{
A linear regression result, similar to summary of a model
}
\description{
Runs a linear regression using bootstrapped estimates to create the covarience matrix.
}
