% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_bias.R
\name{check_bias}
\alias{check_bias}
\title{Checks omitted variable bias in a variable between an initial regression and a longer one}
\usage{
check_bias(
  df,
  response,
  predictor_to_check,
  other_predictors,
  omitted_variables
)
}
\arguments{
\item{df}{A dataframe R will pull the variables from to run the regressions}

\item{response}{The dependent variable of the regressions, as a string}

\item{predictor_to_check}{The independent variable to be checked for bias}

\item{other_predictors}{A vector of strings containing each independent variable used in the initial model
and the long model}

\item{omitted_variables}{A vector of strings containing each independent variable used in the long model only,
also known as the omitted variables}
}
\value{
A sentence in the form of a vector with each omitted variable and the change in the estimate
of the biasvar in question
}
\description{
Checks omitted variable bias in a variable between an initial regression and a longer one
}
