% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_bias.R
\name{check_bias_all}
\alias{check_bias_all}
\title{Checks omitted variable bias in a variable between an initial regression and a longer one}
\usage{
check_bias_all(df, model, ...)
}
\arguments{
\item{df}{A dataframe}

\item{model}{A model object of class "lm" or a valid formula.}

\item{...}{Any number of (unquoted) variables to omitted for checking bias of
the target predictor}
}
\value{
A tibble with each omitted variable and the resulting change in the
estimate of the target predictor in question
}
\description{
Checks omitted variable bias in a variable between an initial regression and a longer one
}
\examples{

check_bias(mtcars, mpg ~ cyl + hp, hp, am)

}
