% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_predictors.R
\name{check_predictors}
\alias{check_predictors}
\title{Starting with a base linear regression, check_predictors takes turns adding a variables from a list of
variables to it one at a time, calculating the change in RSquared Adjusted for each variable addition to
to the base regression, and showing the user if the variable was good or bad for the RSquared Adjusted}
\usage{
check_predictors(df, model, ...)
}
\arguments{
\item{df}{A dataframe}

\item{model}{A model object of class "lm" or a valid formula.}

\item{...}{Any number of (unquoted) variables to be checked as possible additions to the model}
}
\value{
A tibble with each tested variable, whether it was good or bad for
Adjusted RSquared, and the exact change in Adjusted RSquared.
}
\description{
Starting with a base linear regression, check_predictors takes turns adding a variables from a list of
variables to it one at a time, calculating the change in RSquared Adjusted for each variable addition to
to the base regression, and showing the user if the variable was good or bad for the RSquared Adjusted
}
\examples{

check_predictors(mtcars, mpg ~ cyl, hp, am)

}
