% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R1.R
\name{checkvars}
\alias{checkvars}
\title{Starting with a base linear regression, checkvars takes turns adding a variables from a list of
variables to it one at a time, calculating the change in RSquared Adjusted for each variable addition to
to the base regression, and showing the user if the variable was good or bad for the RSquared Adjusted}
\usage{
checkvars(df, output, startersvec, inputsvec)
}
\arguments{
\item{df}{A dataframe R will pull the variables from to run the regressions}

\item{output}{The dependent variable of the regressions, as a string}

\item{startersvec}{A vector of strings containing each independent variable used in the initial model}

\item{inputsvec}{A vector of strings containing each independent variable to be tested with the initial
model}
}
\value{
A sentence in the form of a vector with each tested variable, whether it was good or bad for
Adjusted RSquared, and the exact change in Adjusted RSquared, rounded to 6 decimal places
}
\description{
Starting with a base linear regression, checkvars takes turns adding a variables from a list of
variables to it one at a time, calculating the change in RSquared Adjusted for each variable addition to
to the base regression, and showing the user if the variable was good or bad for the RSquared Adjusted
}
