% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered_errors.R
\name{clustered_errors}
\alias{clustered_errors}
\title{Runs a simple or multiple regression with clustered standard errors}
\usage{
clustered_errors(df, response, predictors, cluster_predictor)
}
\arguments{
\item{df}{A dataframe R will pull the variables from to run the regression}

\item{response}{The dependent variable of the regression, as a string}

\item{predictors}{A vector of strings containing each independent variable used in the model}

\item{cluster_predictor}{A variable to cluster the standard errors by.}
}
\description{
Runs a simple or multiple regression with clustered standard errors
}
