% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors_lm.R
\name{clustered_errors_lm}
\alias{clustered_errors_lm}
\title{Computes a linear regression with clustered standard errors.}
\usage{
clustered_errors_lm(df, model, cluster_predictor)
}
\arguments{
\item{df}{A dataframe R will pull the variables from to run the regression}

\item{model}{A model object of class "lm" or a valid formula.}

\item{cluster_predictor}{An (unquoted) variable to cluster the standard errors by.}
}
\value{
A data frame of model estimates.
}
\description{
Computes a linear regression with clustered standard errors.
}
\examples{

clustered_errors_lm(mtcars, mpg ~ hp + wt, cyl)

}
