% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R3.R
\name{clusterSElm}
\alias{clusterSElm}
\title{Runs a simple or multiple regression with clustered standard errors}
\usage{
clusterSElm(df, output, startersvec, clustervar)
}
\arguments{
\item{df}{A dataframe R will pull the variables from to run the regression}

\item{output}{The dependent variable of the regression, as a string}

\item{startersvec}{A vector of strings containing each independent variable used in the model}

\item{clustervar}{A variable to cluster the standard errors by.}
}
\description{
Runs a simple or multiple regression with clustered standard errors
}
