% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrected_errors.R
\name{corrected_errors}
\alias{corrected_errors}
\title{Runs a simple or multiple regression, automatically correcting for possible heteroskedasticity in
the standard errors}
\usage{
corrected_errors(df, response, predictors, type = "HC")
}
\arguments{
\item{df}{A dataframe R will pull the variables from to run the regression}

\item{response}{The dependent variable of the regression, as a string}

\item{predictors}{A vector of strings containing each independent variable used in the model}

\item{type}{An optional term for specifiying the specific estimator wanted, see vcovHC for options}
}
\value{
Returns summarized output of a linear regression model with corrected standard errors
}
\description{
Runs a simple or multiple regression, automatically correcting for possible heteroskedasticity in
the standard errors
}
