% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors_lm.R
\name{corrected_errors_lm}
\alias{corrected_errors_lm}
\title{Computes a linear regression with standard errors corrected for
heteroskedasticity,  using the}
\usage{
corrected_errors_lm(df, model, cluster_predictor, type = "HC")
}
\arguments{
\item{df}{A dataframe R will pull the variables from to run the regression}

\item{model}{A model object of class "lm" or a valid formula.}

\item{cluster_predictor}{An (unquoted) variable to cluster the standard errors by.}

\item{type}{An optional term for specifiying the estimator method, see vcovHC for options}
}
\value{
A data frame of model estimates.
}
\description{
Computes a linear regression with standard errors corrected for
heteroskedasticity,  using the
}
\examples{

corrected_errors_lm(mtcars, mpg ~ hp + wt, cyl)

}
