% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R3.R
\name{correctedSElm}
\alias{correctedSElm}
\title{Runs a simple or multiple regression, automatically correcting for possible heteroskedasticity in
the standard errors}
\usage{
correctedSElm(df, output, startersvec, type = "HC1")
}
\arguments{
\item{df}{A dataframe R will pull the variables from to run the regression}

\item{output}{The dependent variable of the regression, as a string}

\item{startersvec}{A vector of strings containing each independent variable used in the model}

\item{type}{An optional term for specifiying the specific estimator wanted, see vcovHC for options}
}
\value{
Returns summarized output of a linear regression model with corrected standard errors
}
\description{
Runs a simple or multiple regression, automatically correcting for possible heteroskedasticity in
the standard errors
}
