% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{model_to_formula}
\alias{model_to_formula}
\title{Extracts formula call from an lm object}
\usage{
model_to_formula(model)
}
\arguments{
\item{model}{An object of class "lm"}
}
\value{
An object of class "formula"
}
\description{
Extracts formula call from an lm object
}
