% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R1.R
\name{OmVarBias}
\alias{OmVarBias}
\title{Checks omitted variable bias in a variable between an initial regression and a longer one}
\usage{
OmVarBias(df, response, biasvar, omvars, predictors)
}
\arguments{
\item{df}{A dataframe}

\item{response}{The dependent variable of the regressions, as a string}

\item{biasvar}{The independent variable to be checked for bias}

\item{omvars}{A vector of strings containing each independent variable used in the long model only,
also known as the omitted variables}

\item{predictors}{A vector of strings containing each independent variable used in the initial model
and the long model}
}
\value{
A sentence in the form of a vector with each omitted variable and the change in the estimate
of the biasvar in question
}
\description{
Checks omitted variable bias in a variable between an initial regression and a longer one
}
