% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_plots.R
\name{residual_boxplots}
\alias{residual_boxplots}
\title{Creates scatter plot of residuals and fitted values
(and optionally also selected variables).}
\usage{
residual_boxplots(df, model, ...)
}
\arguments{
\item{df}{A data frame}

\item{model}{A model object or valid formula}

\item{...}{Any number of (unquoted) variable names to plot the residuals
against.}
}
\value{
A boxplot of the relationship between the supplied variable(s) and the residuals
}
\description{
Creates scatter plot of residuals and fitted values
(and optionally also selected variables).
}
\examples{

mtcars \%>\%
    residual_boxplots(mpg ~ hp + cyl, am)

}
