% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_plots.R
\name{residual_histogram}
\alias{residual_histogram}
\title{Plots the histogram of regression residuals to check for normality.
Includes results of Shapiro-Wilk test in visualization.}
\usage{
residual_histogram(df, model)
}
\arguments{
\item{df}{A dataframe R will pull the variables from to run the regression}

\item{model}{A model object or valid formula.}
}
\value{
A density histogram of the residuals
}
\description{
Plots the histogram of regression residuals to check for normality.
Includes results of Shapiro-Wilk test in visualization.
}
\examples{

mtcars \%>\%
    residual_histogram(hp ~ cyl)

my_lm <- lm(hp ~ cyl, data = mtcars)

mtcars \%>\%
    residual_histogram(my_lm)

}
