% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_plots.R
\name{residual_scatterplots}
\alias{residual_scatterplots}
\title{Creates scatter plot of residuals and fitted values (or selected variables).}
\usage{
residual_scatterplots(df, model, ...)
}
\arguments{
\item{df}{A dataframe R will pull the variables from to run the regression}

\item{model}{A model object}

\item{...}{Any number of (unquoted) variable names to plot the residuals
against. If none are supplied, the response variable will be used.}

\item{formula}{The formula for the desired regression.}
}
\value{
A scatter plot of the relationship between the supplied variable(s) and the residuals
}
\description{
Creates scatter plot of residuals and fitted values (or selected variables).
}
\examples{

mtcars \%>\%
    plot_residuals(mpg ~ hp + cyl)

my_lm <- lm(mpg ~ hp + cyl, data = mtcars)

mtcars \%>\%
    plot_residuals(model = my_lm)

}
