% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_visual_residual.R
\name{simple_visual_residual}
\alias{simple_visual_residual}
\title{Plots the relationship of an dependent variable in a simple regression to the residuals of that
regression}
\usage{
simple_visual_residual(df, response, predictor)
}
\arguments{
\item{df}{A dataframe R will pull the variables from to run the regression}

\item{response}{The dependent variable of the regression, as a string}

\item{predictor}{The single dependent variable of the regression, as a string}
}
\value{
A linear plot of the relationship between the dependent variable and the residuals
}
\description{
Plots the relationship of an dependent variable in a simple regression to the residuals of that
regression
}
