% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2.R
\name{simplelmvis}
\alias{simplelmvis}
\title{Plots the relationship of an dependent variable in a simple regression to the residuals of that
regression}
\usage{
simplelmvis(df, output, xvar)
}
\arguments{
\item{df}{A dataframe R will pull the variables from to run the regression}

\item{output}{The dependent variable of the regression, as a string}

\item{xvar}{The single dependent variable of the regression, as a string}
}
\value{
A linear plot of the relationship between the dependent variable and the residuals
}
\description{
Plots the relationship of an dependent variable in a simple regression to the residuals of that
regression
}
