% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_lm.R
\name{tidy_lm}
\alias{tidy_lm}
\title{A simple wrapper on \code{lm()} that takes input in a pipe-friendly order
and gives output tidied by \code{broom::tidy()}}
\usage{
tidy_lm(df, model, ...)
}
\arguments{
\item{df}{A dataframe R will pull the variables from to run the regression}

\item{model}{A model object of class "lm" or a valid formula.}

\item{...}{Further arguments to \code{lm()}}
}
\value{
A data frame of model information
}
\description{
A simple wrapper on \code{lm()} that takes input in a pipe-friendly order
and gives output tidied by \code{broom::tidy()}
}
\examples{

tidy_lm(mtcars, mpg ~ cyl + am + hp, ...)

}
