% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_predictors.R
\name{try_one_predictor}
\alias{try_one_predictor}
\title{Helper function that tries adding one predictor to a model and checks Adjusted R Squared.}
\usage{
try_one_predictor(df, partial_formula, new_predictor, crit_val)
}
\arguments{
\item{df}{A dataframe}

\item{partial_formula}{String representing the formula of the original model.}

\item{new_predictor}{Name of new independent variable to check.}

\item{crit_val}{The Adjusted R-Squared value of the original model.}
}
\value{
A sentence stating whether the new variable was good or bad for the Adjusted RSquared,
and the exact change in Adjusted RSquared, rounded to 6 decimal places
}
\description{
Helper function that tries adding one predictor to a model and checks Adjusted R Squared.
}
