% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2.R
\name{visresid}
\alias{visresid}
\title{Visualizes the distribution of, simple or multiple, regression residuals, checking for normal distribution}
\usage{
visresid(df, output, startersvec)
}
\arguments{
\item{df}{A dataframe R will pull the variables from to run the regression}

\item{output}{The dependent variable of the regression, as a string}

\item{startersvec}{A vector of strings containing each independent variable used in the model}
}
\value{
A density histogram of the residuals
}
\description{
Visualizes the distribution of, simple or multiple, regression residuals, checking for normal distribution
}
