% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_residual.R
\name{visual_residual}
\alias{visual_residual}
\title{Visualizes the distribution of, simple or multiple, regression residuals, checking for normal distribution}
\usage{
visual_residual(df, response, predictors)
}
\arguments{
\item{df}{A dataframe R will pull the variables from to run the regression}

\item{response}{The dependent variable of the regression, as a string}

\item{predictors}{A vector of strings containing each independent variable used in the model}
}
\value{
A density histogram of the residuals
}
\description{
Visualizes the distribution of, simple or multiple, regression residuals, checking for normal distribution
}
