% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_predictors.R
\name{check_predictors}
\alias{check_predictors}
\title{Starting with a base linear regression, checkvars takes turns adding a variables from a list of
variables to it one at a time, calculating the change in RSquared Adjusted for each variable addition to
to the base regression, and showing the user if the variable was good or bad for the RSquared Adjusted}
\usage{
check_predictors(df, response, predictors, predictors_to_check)
}
\arguments{
\item{df}{A dataframe}

\item{response}{The dependent variable of the regressions, as a string}

\item{predictors}{A vector of strings containing the independent variables already in the model}

\item{predictors_to_check}{A vector of strings containing the independent variable to be checked
as possible additions to the model}
}
\value{
A sentence in the form of a vector with each tested variable, whether it was good or bad for
Adjusted RSquared, and the exact change in Adjusted RSquared, rounded to 6 decimal places
}
\description{
Starting with a base linear regression, checkvars takes turns adding a variables from a list of
variables to it one at a time, calculating the change in RSquared Adjusted for each variable addition to
to the base regression, and showing the user if the variable was good or bad for the RSquared Adjusted
}
\examples{

response = c("mpg")
explan <- c("am", "hp")
to_try <- c("wt", "gear", "cyl")

check_predictors(mtcars, response, explan, to_try)

}
