% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_plots.R
\name{residual_histogram}
\alias{residual_histogram}
\title{Visualizes the distribution of regression residuals, checking for normal distribution}
\usage{
residual_histogram(df, model)
}
\arguments{
\item{df}{A dataframe R will pull the variables from to run the regression}

\item{model}{A model object or valid formula.}
}
\value{
A density histogram of the residuals
}
\description{
Visualizes the distribution of regression residuals, checking for normal distribution
}
\examples{

mtcars \%>\%
    plot_residual_density(hp ~ cyl)

my_lm <- lm(hp ~ cyl, data = mtcars)

mtcars \%>\%
    plot_residual_density(my_lm)

}
