% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_plots.R
\name{residual_scatterplots}
\alias{residual_scatterplots}
\title{Creates scatter plot of residuals and fitted values
(and optionally also selected variables).}
\usage{
residual_scatterplots(df, model, ...)
}
\arguments{
\item{df}{A data frame}

\item{model}{A model object or valid formula}

\item{...}{Any number of (unquoted) variable names to plot the residuals
against. If none are supplied, only fitted values will be shown.}
}
\value{
A scatter plot of the relationship between the supplied variable(s) and the residuals
}
\description{
Creates scatter plot of residuals and fitted values
(and optionally also selected variables).
}
\examples{

mtcars \%>\%
    residual_scatterplots(mpg ~ hp + cyl, hp, wt)

}
